add executable and more cpplint fixes

+++ src/griddyn/solvers/sundialsInterface.cpp
@@ -11,6 +11,7 @@
  */
 
 #include "griddyn/griddyn-config.h"
+#include "sundialsInterface.h"
 #include "idaInterface.h"
 #include "kinsolInterface.h"
 #ifdef GRIDDYN_ENABLE_CVODE
@@ -391,7 +392,7 @@ namespace solvers {
             sd->nnz = a1->size();
             if (sd->flags[fileCapture_flag]) {
                 if (!sd->jacFile.empty()) {
-                    auto val = static_cast<long int>(sd->get("nliterations"));
+                    auto val = static_cast<std::uint32_t>(sd->get("nliterations"));
                     writeArray(time, 1, val, sd->mode.offsetIndex, *a1, sd->jacFile);
                 }
             }
